requirements.txt
reflex==0.8.24
any-llm-sdk[gemini]>=1.2.0
neo4j>=5.0.0
pydantic>=2.0.0
fonttools>=4.0.0
pynacl>=1.5.0
pikepdf>=8.0.0
pillow>=10.0.0
coverage>=7.0.0
playwright>=1.49.0
rich>=13.7.0
diagrams>=0.23.3
graphviz>=0.20.3
pydot>=4.0.1
pydeps>=3.0.1
pyan3>=1.2.0
pylint>=3.3.9
ruff>=0.6.9
black>=24.8.0
isort>=5.13.2
mypy>=1.13.0
flake8>=7.1.1
pyflakes>=3.2.0
pycodestyle>=2.12.1
pydocstyle>=6.3.0
pyright>=1.1.388
vulture>=2.14
bandit>=1.7.9
boltons~=21.0
click-option-group~=0.5
codespell>=2.2.6
check-jsonschema>=0.28.0
detect-secrets>=1.4.0
pip-audit>=2.7.3
safety>=3.2.0
pip-check>=3.2.1
pipdeptree>=2.30.0
pydoclint>=0.8.3
pyupgrade>=3.15.2
dodgy>=0.2.1
eradicate>=2.3.0
deptry>=0.23.0
pycln>=2.5.0
autoflake>=2.3.1
refurb>=2.2.0
radon>=6.0.1
xenon>=0.9.1
lizard>=1.17.10
mccabe>=0.7.0
interrogate>=1.7.0
exceptiongroup~=1.2.0
glom~=22.1
mcp==1.23.3
opentelemetry-api~=1.37.0
opentelemetry-exporter-otlp-proto-http~=1.37.0
opentelemetry-instrumentation-requests~=0.58b0
opentelemetry-sdk~=1.37.0
peewee~=3.14
pyre-check>=0.9.20
wcmatch~=8.3
